Add Windows 10 installer example pipeline
This adds an example pipeline for kubernetes and okd, which creates a
base DataVolume and performs a basic install of Windows 10. On okd the
windows10 template of common-templates is used to create a new template
after the installation. The new template is prepopulated with the created
base DataVolume and allows to create new VMs which can be used without
installation. An example autounattend.xml for initial setup of the VM is
included in the pipeline's ConfigMap windows10-autounatttend. For basic
setup of a preinstalled VMs an example unattend.xml is included in the
pipeline's ConfigMap windows10-unattend.

Example PipelineRun for kubernetes:
```
apiVersion: tekton.dev/v1beta1
kind: PipelineRun
metadata:
  generateName: windows10-installer-run-
spec:
  params:
  - name: winImageDownloadURL
    value: INSERT_WINDOWS_ISO_URL
  pipelineRef:
    name: windows10-installer
  serviceAccountNames:
    - taskName: create-source-dv
      serviceAccountName: create-datavolume-from-manifest-task
    - taskName: create-base-dv
      serviceAccountName: create-datavolume-from-manifest-task
    - taskName: create-vm-from-manifest
      serviceAccountName: create-vm-from-manifest-task
    - taskName: wait-for-vmi-status
      serviceAccountName: wait-for-vmi-status-task
    - taskName: cleanup-vm
      serviceAccountName: cleanup-vm-task
status: {}
```

Example PipelineRun for okd:
```
apiVersion: tekton.dev/v1beta1
kind: PipelineRun
metadata:
  generateName: windows10-installer-run-
spec:
  params:
  - name: winImageDownloadURL
    value: INSERT_WINDOWS_ISO_URL
  pipelineRef:
    name: windows10-installer
  serviceAccountNames:
    - taskName: create-base-dv
      serviceAccountName: create-datavolume-from-manifest-task
    - taskName: copy-template-installer
      serviceAccountName: copy-template-task
    - taskName: modify-vm-template-installer
      serviceAccountName: modify-vm-template-task
    - taskName: create-vm-from-template
      serviceAccountName: create-vm-from-template-task
    - taskName: wait-for-vmi-status
      serviceAccountName: wait-for-vmi-status-task
    - taskName: cleanup-vm
      serviceAccountName: cleanup-vm-task
    - taskName: copy-template-golden
      serviceAccountName: copy-template-task
    - taskName: modify-vm-template-golden
      serviceAccountName: modify-vm-template-task
status: {}
```

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>